% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Utility.R
\name{clear.data}
\alias{clear.data}
\title{Erase all Data for an MN Object}
\usage{
clear.data(mn)
}
\arguments{
\item{mn}{A MetaNet Object.}
}
\value{
Nothing - Evaluated in Place
}
\description{
Erase all Data for an MN Object
}
\examples{
data(dataset1)
data(dataset2)
data(dataset3)
files <- list(dataset1,dataset2,dataset3)
lpv <- list(list(1,2:4,5:7),list(1,2:4,5:7),list(1,2:4,5:7))
mn <- load.data(f=files,lpv=lpv)
map.network(mn,datasets=c(1,2,3),lpv=list(list(1,1:2,1:2),list(1,1:2,1),list(1,1:2,2:3)))
clear.data(mn)
}

