% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Utility.R
\name{get.data}
\alias{get.data}
\title{Get a Data Set from a MN Object}
\usage{
get.data(mn, index)
}
\arguments{
\item{mn}{A MetaNet Object.}

\item{index}{An Integer}
}
\value{
A Data Frame
}
\description{
Get a Data Set from a MN Object
}
\examples{
data(dataset1)
data(dataset2)
data(dataset3)
files <- list(dataset1,dataset2,dataset3)
lpv <- list(list(1,2:4,5:7),list(1,2:4,5:7),list(1,2:4,5:7))
mn <- load.data(f=files,lpv=lpv)
map.network(mn,datasets=c(1,2,3),lpv=list(list(1,1:2,1:2),list(1,1:2,1),list(1,1:2,2:3)))
set.2 <- get.data(mn,2)
}

