% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IO.R
\name{load.data}
\alias{load.data}
\title{Load Annotated Data.}
\usage{
load.data(f, lpv, sep = "\\t", quote = "\\"", header = T)
}
\arguments{
\item{f}{A vector of file paths.}

\item{lpv}{A list of LPV mappings.}

\item{sep}{Any additional params}

\item{quote}{The qUOTE}

\item{header}{The header}
}
\value{
A MetaNet object.
}
\description{
Load Annotated Data.
}
\examples{
data(dataset1)
data(dataset2)
data(dataset3)
files <- list(dataset1,dataset2,dataset3)
lpv <- list(list(1,2:4,5:7),list(1,2:4,5:7),list(1,2:4,5:7))
mn <- load.data(f=files,lpv=lpv)
}

