% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Methods.R
\name{map.network}
\alias{map.network}
\title{Map Networks from an MN Object}
\usage{
map.network(mn, datasets = c(), lpv = list(), multigraph = F,
  standard = F, in.place = T, add = NA)
}
\arguments{
\item{mn}{A MetaNet Object.}

\item{datasets}{A Vector}

\item{lpv}{A list - LPV Mapc}

\item{multigraph}{A Boolean}

\item{standard}{A Boolean}

\item{in.place}{A Boolean}

\item{add}{A Vector}
}
\value{
An MN Object if in.place=F, else Nothing - eval in place.
}
\description{
Map Networks from an MN Object
}
\examples{
data(dataset1)
data(dataset2)
data(dataset3)
files <- list(dataset1,dataset2,dataset3)
lpv <- list(list(1,2:4,5:7),list(1,2:4,5:7),list(1,2:4,5:7))
mn <- load.data(f=files,lpv=lpv)
map.network(mn,datasets=c(1,2,3),lpv=list(list(1,1:2,1:2),list(1,1:2,1),list(1,1:2,2:3)))
}

