% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Utility.R
\name{set.LPV}
\alias{set.LPV}
\title{Set new LPV mapping for an MN Object's Data Set}
\usage{
set.LPV(mn, index, label.cols = NA, path.cols = NA, value.cols = NA)
}
\arguments{
\item{mn}{A MetaNet Object.}

\item{index}{A Number}

\item{label.cols}{A Vector}

\item{path.cols}{A Vector}

\item{value.cols}{A Vector}
}
\value{
Nothing - Evaluated in Place
}
\description{
Set new LPV mapping for an MN Object's Data Set
}
\examples{
data(dataset1)
data(dataset2)
data(dataset3)
files <- list(dataset1,dataset2,dataset3)
lpv <- list(list(1,2:4,5:7),list(1,2:4,5:7),list(1,2:4,5:7))
mn <- load.data(f=files,lpv=lpv)
map.network(mn,datasets=c(1,2,3),lpv=list(list(1,1:2,1:2),list(1,1:2,1),list(1,1:2,2:3)))
set.LPV(mn,index=2,label.cols=1,path.cols=3:5,value.cols=6:7)
}

