% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Utility.R
\name{set.metadata}
\alias{set.metadata}
\title{Set MetaData for an MN Object's Data Set}
\usage{
set.metadata(mn, pass.in, value)
}
\arguments{
\item{mn}{A MetaNet Object.}

\item{pass.in}{A string}

\item{value}{A string}
}
\value{
Nothing - Evaluated in Place
}
\description{
Set MetaData for an MN Object's Data Set
}
\examples{
data(dataset1)
data(dataset2)
data(dataset3)
files <- list(dataset1,dataset2,dataset3)
lpv <- list(list(1,2:4,5:7),list(1,2:4,5:7),list(1,2:4,5:7))
mn <- load.data(f=files,lpv=lpv)
map.network(mn,datasets=c(1,2,3),lpv=list(list(1,1:2,1:2),list(1,1:2,1),list(1,1:2,2:3)))
set.metadata(mn,'location', 'bahamas')
}

